feat: Cache transformed selection sets
This avoids a potential performance trap when relying heavily fragment
composition. When traversing a selection set to generate its type
declaration, we fully resolve all encountered fields. When the field is
a fragment, we recursively resolve it, even if we have already done so
earlier in the traversal.

For certain composition patterns, this means we resolve the same
fragment a large number of times, potentially incurring significant CPU
cost.

By adding a resolution cache, we only resolve a given selection set (a
fragment or a selection set of fields) once per GQL type. This is safe,
as selecting the same thing on the same GQL type always yields the same
type information.